requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
fastapi
uvicorn
langchain-google-community
google-auth-oauthlab